// apps/frontend/src/api/api.ts
export interface GroupInfoDto {
  /** 群号 */
  gc: number;
  /** 群名称 */
  gn: string;
  /** 角色 */
  role: "creator" | "manager" | "member";
  /** 角色名称: 群主/管理员/群员 */
  roleName: string;
}
// ...
export interface IQQLoginCookies {
  pt2gguin?: string;
  pt4_token?: string;
  RK?: string;
  p_skey?: string;
  p_uin?: string;
  pt_recent_uins?: string;
  ptcz?: string;
  ptnick_?: string;
  skey?: string;
  superkey?: string;
  supertoken?: string;
  superuin?: string;
  uin?: string;
}